Refactor for maven, add single cmdline interface, update Wheeljack
We refactored for maven and use a single pom.xml file for building the
jar.  We then added an interface that allows a single entry point for
the application and selects from a series of commands.  Commands should
use the annotation CommandLineTool and optionally
CommandLineDescription, the entry point will then search for the correct
class.

Wheeljack had a minor update that allows a major new feature, we can now
filter potential alleles by allele fraction.  Note that this is the
allele fraction for the fully genotyped allele, so you may want to
account for loss of allele counts due to sequence error.  For example,
if you are planning to get alleles that are at 5% you may want to set
the filter to 4.5% to account for some loss due to sequence errors in
surrounding bases.